filter sscs5 benefit types
