+++ steps/start/benefit-type/BenefitType.js
@@ -4,14 +4,13 @@ const { form, text } = require('@hmcts/one-per-page/forms');
 const { answer } = require('@hmcts/one-per-page/checkYourAnswers');
 const {
   splitBenefitType,
-  getBenefitCode,
-  isFeatureFlagEnabled
+  getBenefitCode
 } = require('utils/stringUtils');
 const sections = require('steps/check-your-appeal/sections');
 const Joi = require('joi');
 const paths = require('paths');
 const benefitTypes = require('steps/start/benefit-type/types');
-const { isIba } = require('utils/benefitTypeUtils');
+const { isIba, getAllowedBenefitTypes } = require('utils/benefitTypeUtils');
 const config = require('config');
 const i18next = require('i18next');
 
@@ -62,42 +61,7 @@ class BenefitType extends SaveToDraftStore {
   }
 
   getAllowedTypes() {
-    const allowedTypes = [
-      benefitTypes.personalIndependencePayment,
-      benefitTypes.employmentAndSupportAllowance,
-      benefitTypes.universalCredit,
-      benefitTypes.infectedBloodCompensation
-    ];
-
-    const featureFlags = [
-      { flag: 'allowDLA', benefit: benefitTypes.disabilityLivingAllowance },
-      { flag: 'allowCA', benefit: benefitTypes.carersAllowance },
-      { flag: 'allowAA', benefit: benefitTypes.attendanceAllowance },
-      { flag: 'allowBB', benefit: benefitTypes.bereavementBenefit },
-      {
-        flag: 'allowIIDB',
-        benefit: benefitTypes.industrialInjuriesDisablement
-      },
-      { flag: 'allowJSA', benefit: benefitTypes.jobseekersAllowance },
-      { flag: 'allowSF', benefit: benefitTypes.socialFund },
-      { flag: 'allowMA', benefit: benefitTypes.maternityAllowance },
-      { flag: 'allowIS', benefit: benefitTypes.incomeSupport },
-      {
-        flag: 'allowBSPS',
-        benefit: benefitTypes.bereavementSupportPaymentScheme
-      },
-      { flag: 'allowIDB', benefit: benefitTypes.industrialDeathBenefit },
-      { flag: 'allowPC', benefit: benefitTypes.pensionCredit },
-      { flag: 'allowRP', benefit: benefitTypes.retirementPension }
-    ];
-
-    allowedTypes.push(
-      ...featureFlags
-        .filter(f => isFeatureFlagEnabled(f.flag))
-        .map(f => f.benefit)
-    );
-
-    return allowedTypes;
+    return getAllowedBenefitTypes();
   }
 
   next() {
